+++ GiraffeRender/GIRTextureLoader+HDR.m
@@ -15,7 +15,7 @@
 
 @implementation GIRHDRLoader
 
-+ (float*)loadHDR:(NSString * _Nonnull)hdrPath width:(NSInteger * _Nonnull)width height:(NSInteger * _Nonnull)height numComponents:(NSInteger * _Nonnull)numOfConmonents
++ (float*)loadHDR:(NSString * _Nonnull)hdrPath width:(int32_t * _Nonnull)width height:(int32_t * _Nonnull)height numComponents:(int32_t * _Nonnull)numOfConmonents
 {
     NSString *path = [[NSBundle mainBundle] pathForResource:@"Art.scnassets/HDR/newport_loft" ofType:@"hdr"];
     float *data = stbi_loadf([path UTF8String], width, height, numOfConmonents, 4); // TODO: need to free resource
